For Actor.java --> Reorganized setName() and getName(); created getEnemyName(String) and String e_name for the enemy classes. Changed all protected types to public types. Changed int types to double types (as needed).
For Enemy.java -->
Made minor logical adjustments. setSpeed() is still static and sets it to 1.0. The other methods were changed to abstract methods. The getDamage() method was removed since it is calculated in battle.

For Base_Enemy.java -->
This is the first time this class is being uploaded to the project through InteliJ. If there are any problems that you see with it, let me know.